LPS-162411 Change structure of ActivityClaimPanel
