+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/extra/remote-app/src/routes/MDFClaimForm/components/ActivityClaimPanel/index.ts
@@ -9,4 +9,4 @@
  * distribution rights of the Software.
  */
 
-export {default} from './BudgetModal';
+export {default} from './ActivityClaimPanel';
